Set default value while getting environments_list option for tests
