--- spyder_env_manager/tests/test_container.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# ----------------------------------------------------------------------------
-# Copyright © 2022, Spyder Development Team and spyder-env-manager contributors
-#
-# Licensed under the terms of the MIT license
-# ----------------------------------------------------------------------------
-"""
-Spyder Env Manager container tests.
-"""
