+++ spyder_env_manager/tests/test_plugin.py
@@ -7,14 +7,3 @@
 """
 Spyder Env Manager plugin tests.
 """
-
-# Local imports
-from spyder_env_manager.spyder.widgets.main_widget import SpyderEnvManagerWidget
-
-
-def test_env_manager(qtbot):
-    """Create plugin widget and show it."""
-    widget = SpyderEnvManagerWidget(None)
-    widget.setup()
-    qtbot.addWidget(widget)
-    widget.show()
